// src/schema/types.generated.ts
export type Maybe<T> = T | null | undefined;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type ARecord = {
  __typename?: 'ARecord';
  address: Scalars['String']['output'];
  ttl: Scalars['Int']['output'];
};

export type CName = {
  __typename?: 'CName';
  hostname: Scalars['String']['output'];
};

export type DnsQueryResult = ARecord | CName;

export type Query = {
  __typename?: 'Query';
  dnsLookup?: Maybe<Array<DnsQueryResult>>;
  myIp: Scalars['String']['output'];
};


export type QuerydnsLookupArgs = {
  address: Scalars['String']['input'];
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversUnionTypes<RefType extends Record<string, unknown>> = {
  DnsQueryResult: ( ARecord & { __typename: 'ARecord' } ) | ( CName & { __typename: 'CName' } );
};


/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  ARecord: ResolverTypeWrapper<ARecord>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
  Int: ResolverTypeWrapper<Scalars['Int']['output']>;
  CName: ResolverTypeWrapper<CName>;
  DnsQueryResult: ResolverTypeWrapper<ResolversUnionTypes<ResolversTypes>['DnsQueryResult']>;
  Query: ResolverTypeWrapper<{}>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  ARecord: ARecord;
  String: Scalars['String']['output'];
  Int: Scalars['Int']['output'];
  CName: CName;
  DnsQueryResult: ResolversUnionTypes<ResolversParentTypes>['DnsQueryResult'];
  Query: {};
  Boolean: Scalars['Boolean']['output'];
};
// ...
export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  dnsLookup?: Resolver<Maybe<Array<ResolversTypes['DnsQueryResult']>>, ParentType, ContextType, RequireFields<QuerydnsLookupArgs, 'address'>>;
  myIp?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
};